start merging concurrent and normal connections
