+++ cpp/bindings.cpp
@@ -1,7 +1,6 @@
 #include "bindings.h"
+#include "ConnectionPool.h"
 #include "JSIHelper.h"
-#include "concurrency/ConcurrentSQLiteBridge.h"
-#include "concurrency/ConnectionPool.h"
 #include "logs.h"
 #include "macros.h"
 #include "sqlbatchexecutor.h"
@@ -34,10 +33,7 @@ int init_powersync_sqlite_plugin() {
   return result;
 }
 
-void clearState() {
-  closeAllConcurrentDBs();
-  sqliteCloseAll();
-}
+void clearState() { sqliteCloseAll(); }
 
 /**
  * Callback handler for SQLite table updates
@@ -102,8 +98,10 @@ void install(jsi::Runtime &rt,
   invoker = jsCallInvoker;
   runtime = &rt;
 
+  // Any DBs opened after this call will have PowerSync SQLite extension loaded
+  init_powersync_sqlite_plugin();
+
   auto open = HOSTFN("open", 2) {
-    init_powersync_sqlite_plugin();
 
     if (count == 0) {
       throw jsi::JSError(
@@ -128,46 +126,7 @@ void install(jsi::Runtime &rt,
       tempDocPath = tempDocPath + "/" + args[1].asString(rt).utf8(rt);
     }
 
-    sqlite3 *db;
-    SQLiteOPResult result = sqliteOpenDb(dbName, tempDocPath, &db);
-
-    if (result.type == SQLiteError) {
-      throw jsi::JSError(rt, result.errorMessage.c_str());
-    }
-
-    // Register for update hooks
-    sqlite3_update_hook(db, updateTableHandler, getDBName(db));
-
-    return {};
-  });
-
-  auto openConcurrent = HOSTFN("openConcurrent", 2) {
-    init_powersync_sqlite_plugin();
-
-    if (count == 0) {
-      throw jsi::JSError(
-          rt, "[react-native-quick-sqlite][open] database name is required");
-    }
-
-    if (!args[0].isString()) {
-      throw jsi::JSError(
-          rt,
-          "[react-native-quick-sqlite][open] database name must be a string");
-    }
-
-    string dbName = args[0].asString(rt).utf8(rt);
-    string tempDocPath = string(docPathStr);
-
-    if (count > 1 && !args[1].isUndefined() && !args[1].isNull()) {
-      if (!args[1].isString()) {
-        throw jsi::JSError(rt, "[react-native-quick-sqlite][open] database "
-                               "location must be a string");
-      }
-
-      tempDocPath = tempDocPath + "/" + args[1].asString(rt).utf8(rt);
-    }
-
-    auto result = sqliteOpenDBConcurrent(
+    auto result = sqliteOpenDb(
         dbName, tempDocPath, &contextLockAvailableHandler, &updateTableHandler);
     if (result.type == SQLiteError) {
       throw jsi::JSError(rt, result.errorMessage.c_str());
@@ -176,6 +135,7 @@ void install(jsi::Runtime &rt,
     return {};
   });
 
+  // TODO
   auto attach = HOSTFN("attach", 4) {
     if (count < 3) {
       throw jsi::JSError(
@@ -211,6 +171,7 @@ void install(jsi::Runtime &rt,
     return {};
   });
 
+  // TODO
   auto detach = HOSTFN("detach", 2) {
     if (count < 2) {
       throw jsi::JSError(
@@ -236,29 +197,6 @@ void install(jsi::Runtime &rt,
 
   auto close = HOSTFN("close", 1) {
     if (count == 0) {
-      throw jsi::JSError(
-          rt, "[react-native-quick-sqlite][close] database name is required");
-    }
-
-    if (!args[0].isString()) {
-      throw jsi::JSError(
-          rt,
-          "[react-native-quick-sqlite][close] database name must be a string");
-    }
-
-    string dbName = args[0].asString(rt).utf8(rt);
-
-    SQLiteOPResult result = sqliteCloseDb(dbName);
-
-    if (result.type == SQLiteError) {
-      throw jsi::JSError(rt, result.errorMessage.c_str());
-    }
-
-    return {};
-  });
-
-  auto closeConcurrent = HOSTFN("closeConcurrent", 1) {
-    if (count == 0) {
       throw jsi::JSError(rt, "[react-native-quick-sqlite][closeConcurrent] "
                              "database name is required");
     }
@@ -270,7 +208,7 @@ void install(jsi::Runtime &rt,
 
     string dbName = args[0].asString(rt).utf8(rt);
 
-    SQLiteOPResult result = sqliteCloseDBConcurrent(dbName);
+    SQLiteOPResult result = sqliteCloseDb(dbName);
 
     if (result.type == SQLiteError) {
       throw jsi::JSError(rt, result.errorMessage.c_str());
@@ -310,95 +248,6 @@ void install(jsi::Runtime &rt,
     }
 
     return {};
-  });
-
-  auto execute = HOSTFN("execute", 3) {
-    const string dbName = args[0].asString(rt).utf8(rt);
-    const string query = args[1].asString(rt).utf8(rt);
-    vector<QuickValue> params;
-    if (count == 3) {
-      const jsi::Value &originalParams = args[2];
-      jsiQueryArgumentsToSequelParam(rt, originalParams, &params);
-    }
-
-    vector<map<string, QuickValue>> results;
-    vector<QuickColumnMetadata> metadata;
-
-    // Converting results into a JSI Response
-    try {
-      auto status = sqliteExecute(dbName, query, &params, &results, &metadata);
-
-      if (status.type == SQLiteError) {
-        //        throw std::runtime_error(status.errorMessage);
-        throw jsi::JSError(rt, status.errorMessage);
-        //        return {};
-      }
-
-      auto jsiResult =
-          createSequelQueryExecutionResult(rt, status, &results, &metadata);
-      return jsiResult;
-    } catch (std::exception &e) {
-      throw jsi::JSError(rt, e.what());
-    }
-  });
-
-  auto executeAsync = HOSTFN("executeAsync", 3) {
-    if (count < 3) {
-      throw jsi::JSError(rt, "[react-native-quick-sqlite][executeAsync] "
-                             "Incorrect arguments for executeAsync");
-    }
-
-    const string dbName = args[0].asString(rt).utf8(rt);
-    const string query = args[1].asString(rt).utf8(rt);
-    const jsi::Value &originalParams = args[2];
-
-    // Converting query parameters inside the javascript caller thread
-    vector<QuickValue> params;
-    jsiQueryArgumentsToSequelParam(rt, originalParams, &params);
-
-    auto promiseCtr = rt.global().getPropertyAsFunction(rt, "Promise");
-    auto promise = promiseCtr.callAsConstructor(rt, HOSTFN("executor", 2) {
-      auto resolve = std::make_shared<jsi::Value>(rt, args[0]);
-      auto reject = std::make_shared<jsi::Value>(rt, args[1]);
-
-      auto task = [&rt, dbName, query,
-                   params = make_shared<vector<QuickValue>>(params), resolve,
-                   reject]() {
-        try {
-          vector<map<string, QuickValue>> results;
-          vector<QuickColumnMetadata> metadata;
-          auto status =
-              sqliteExecute(dbName, query, params.get(), &results, &metadata);
-          invoker->invokeAsync(
-              [&rt,
-               results = make_shared<vector<map<string, QuickValue>>>(results),
-               metadata = make_shared<vector<QuickColumnMetadata>>(metadata),
-               status_copy = move(status), resolve, reject] {
-                if (status_copy.type == SQLiteOk) {
-                  auto jsiResult = createSequelQueryExecutionResult(
-                      rt, status_copy, results.get(), metadata.get());
-                  resolve->asObject(rt).asFunction(rt).call(rt,
-                                                            move(jsiResult));
-                } else {
-                  auto errorCtr =
-                      rt.global().getPropertyAsFunction(rt, "Error");
-                  auto error = errorCtr.callAsConstructor(
-                      rt, jsi::String::createFromUtf8(
-                              rt, status_copy.errorMessage));
-                  reject->asObject(rt).asFunction(rt).call(rt, error);
-                }
-              });
-        } catch (std::exception &exc) {
-          invoker->invokeAsync([&rt, &exc] { jsi::JSError(rt, exc.what()); });
-        }
-      };
-
-      globalThreadPool->queueWork(task);
-
-      return {};
-    }));
-
-    return promise;
   });
 
   auto executeInContext = HOSTFN("executeInContext", 3) {
@@ -461,34 +310,7 @@ void install(jsi::Runtime &rt,
     return promise;
   });
 
-  // Execute a batch of SQL queries in a transaction
-  // Parameters can be: [[sql: string, arguments: any[] | arguments: any[][] ]]
-  auto executeBatch = HOSTFN("executeBatch", 2) {
-    if (sizeof(args) < 2) {
-      throw jsi::JSError(rt, "[react-native-quick-sqlite][executeBatch] - "
-                             "Incorrect parameter count");
-    }
-
-    const jsi::Value &params = args[1];
-    if (params.isNull() || params.isUndefined()) {
-      throw jsi::JSError(rt, "[react-native-quick-sqlite][executeBatch] - An "
-                             "array of SQL commands or parameters is needed");
-    }
-    const string dbName = args[0].asString(rt).utf8(rt);
-    const jsi::Array &batchParams = params.asObject(rt).asArray(rt);
-    vector<QuickQueryArguments> commands;
-    jsiBatchParametersToQuickArguments(rt, batchParams, &commands);
-
-    auto batchResult = sqliteExecuteBatch(dbName, &commands);
-    if (batchResult.type == SQLiteOk) {
-      auto res = jsi::Object(rt);
-      res.setProperty(rt, "rowsAffected", jsi::Value(batchResult.affectedRows));
-      return move(res);
-    } else {
-      throw jsi::JSError(rt, batchResult.message);
-    }
-  });
-
+  // TODO do this only in a context
   auto executeBatchAsync = HOSTFN("executeBatchAsync", 2) {
     if (sizeof(args) < 2) {
       throw jsi::JSError(rt, "[react-native-quick-sqlite][executeAsyncBatch] "
@@ -609,7 +431,7 @@ void install(jsi::Runtime &rt,
     return promise;
   });
 
-  auto requestConcurrentLock = HOSTFN("requestConcurrentLock", 3) {
+  auto requestLock = HOSTFN("requestLock", 3) {
     if (count < 3) {
       throw jsi::JSError(rt,
                          "[react-native-quick-sqlite][requestConcurrentLock] "
@@ -626,14 +448,14 @@ void install(jsi::Runtime &rt,
     string lockId = args[1].asString(rt).utf8(rt);
     ConcurrentLockType lockType = (ConcurrentLockType)args[2].asNumber();
 
-    auto lockResult = sqliteConcurrentRequestLock(dbName, lockId, lockType);
+    auto lockResult = sqliteRequestLock(dbName, lockId, lockType);
     vector<map<string, QuickValue>> resultsHolder;
     auto jsiResult =
         createSequelQueryExecutionResult(rt, lockResult, &resultsHolder, NULL);
     return jsiResult;
   });
 
-  auto releaseConcurrentLock = HOSTFN("releaseConcurrentLock", 3) {
+  auto releaseLock = HOSTFN("releaseLock", 3) {
     if (count < 2) {
       throw jsi::JSError(rt,
                          "[react-native-quick-sqlite][requestConcurrentLock] "
@@ -649,7 +471,7 @@ void install(jsi::Runtime &rt,
     string dbName = args[0].asString(rt).utf8(rt);
     string lockId = args[1].asString(rt).utf8(rt);
 
-    sqliteConcurrentReleaseLock(dbName, lockId);
+    sqliteReleaseLock(dbName, lockId);
 
     return {};
   });
@@ -657,20 +479,21 @@ void install(jsi::Runtime &rt,
   jsi::Object module = jsi::Object(rt);
 
   module.setProperty(rt, "open", move(open));
-  module.setProperty(rt, "openConcurrent", move(openConcurrent));
-  module.setProperty(rt, "requestConcurrentLock", move(requestConcurrentLock));
-  module.setProperty(rt, "releaseConcurrentLock", move(releaseConcurrentLock));
+  module.setProperty(rt, "requestLock", move(requestLock));
+  module.setProperty(rt, "releaseLock", move(releaseLock));
+  module.setProperty(rt, "executeInContext", move(executeInContext));
   module.setProperty(rt, "close", move(close));
-  module.setProperty(rt, "closeConcurrent", move(closeConcurrent));
+
+  // TODO
   module.setProperty(rt, "attach", move(attach));
+  // TODO
   module.setProperty(rt, "detach", move(detach));
   module.setProperty(rt, "delete", move(remove));
-  module.setProperty(rt, "execute", move(execute));
-  module.setProperty(rt, "executeAsync", move(executeAsync));
-  module.setProperty(rt, "executeBatch", move(executeBatch));
+  // TODO
   module.setProperty(rt, "executeBatchAsync", move(executeBatchAsync));
-  module.setProperty(rt, "executeInContext", move(executeInContext));
+  // TODO
   module.setProperty(rt, "loadFile", move(loadFile));
+  // TODO
   module.setProperty(rt, "loadFileAsync", move(loadFileAsync));
 
   rt.global().setProperty(rt, "__QuickSQLiteProxy", move(module));
